backend/requirements.txt
asgiref==3.5.0
astroid==2.11.5
autopep8==1.6.0
colorama==0.4.4
dill==0.3.5.1
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
importlib-metadata==4.11.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.7.0
Pillow==9.0.1
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
pylint==2.13.9
pytz==2021.3
sqlparse==0.4.2
tk==0.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2021.5
wrapt==1.14.1
zipp==3.7.0
